+++ barapost/barapost-binning.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-__version__ = "4.6.g"
+__version__ = "4.7.a"
 # Year, month, day
-__last_update_date__ = "2020-12-28"
+__last_update_date__ = "2021-01-09"
 
 # |===== Check python interpreter version =====|
 
@@ -782,6 +782,20 @@ with open (logfile_path, 'a') as logfile:
         i += 1
     # end for
 # end with
+
+
+import src.legacy_taxonomy_handling as legacy_taxonomy_handling
+
+# Form path to taxonomy file:
+taxonomy_dir = os.path.join(tax_annot_res_dir, "taxonomy")
+if not os.path.isdir(taxonomy_dir):
+    os.makedirs(taxonomy_dir)
+# end if
+taxonomy_path = os.path.join(taxonomy_dir, "taxonomy.tsv")
+
+# Check if there is legacy taxonomy file and, if so, reformat it to new (TSV) format
+legacy_taxonomy_handling.check_deprecated_taxonomy(tax_annot_res_dir, logfile_path)
+
 
 if n_thr != 1:
     from src.spread_files_equally import spread_files_equally