+++ barapost/src/binning_modules/single_thread_FAST5_binfunc.py
@@ -54,8 +54,8 @@ def bin_fast5_file(f5_path, tax_annot_res_dir, sens,
 
     new_dpath = glob("{}{}*{}*".format(tax_annot_res_dir, os.sep, get_checkstr(f5_path)))[0]
     tsv_res_fpath = get_res_tsv_fpath(new_dpath, logfile_path)
-    resfile_lines = configure_resfile_lines(tsv_res_fpath, sens,
-        os.path.join(tax_annot_res_dir, "taxonomy", "taxonomy"), logfile_path)
+    taxonomy_path = os.path.join(tax_annot_res_dir, "taxonomy", "taxonomy.tsv")
+    resfile_lines = configure_resfile_lines(tsv_res_fpath, sens, taxonomy_path, logfile_path)
 
     # Make filter for quality and length
     QL_filter = get_QL_filter(f5_path, min_qual, min_qlen)