+++ barapost/src/binning_modules/single_thread_FAST5_binfunc_utw.py
@@ -125,8 +125,8 @@ def bin_fast5_file(f5_path, tax_annot_res_dir, sens,
     for tsv_path in index_f5_2_tsv[f5_path].keys():
 
         read_names = index_f5_2_tsv[f5_path][tsv_path]
-        resfile_lines = configure_resfile_lines(tsv_path, sens,
-            os.path.join(tax_annot_res_dir, "taxonomy", "taxonomy"), logfile_path)
+        taxonomy_path = os.path.join(tax_annot_res_dir, "taxonomy", "taxonomy.tsv")
+        resfile_lines = configure_resfile_lines(tsv_path, sens, taxonomy_path, logfile_path)
 
         for read_name in read_names:
             try: