+++ barapost/src/binning_modules/single_thread_QA.py
@@ -101,8 +101,8 @@ def bin_fastqa_file(fq_fa_path, tax_annot_res_dir, sens,
 
     new_dpath = get_curr_res_dpath(fq_fa_path, tax_annot_res_dir)
     tsv_res_fpath = get_res_tsv_fpath(new_dpath, logfile_path)
-    resfile_lines = configure_resfile_lines(tsv_res_fpath, sens,
-        os.path.join(tax_annot_res_dir, "taxonomy", "taxonomy"), logfile_path)
+    taxonomy_path = os.path.join(tax_annot_res_dir, "taxonomy", "taxonomy.tsv")
+    resfile_lines = configure_resfile_lines(tsv_res_fpath, sens, taxonomy_path, logfile_path)
 
     # Configure generator, write function and path to trash file
     if is_fastq(fq_fa_path):